requirements.txt
py-trello ~= 0.16.0
tzlocal ~= 2.0
version-query >= 1.0.5, == 1.*
